--- xwiki-platform-core/xwiki-platform-requiredrights/xwiki-platform-requiredrights-api/src/main/resources/META-INF/components.txt
@@ -1,1 +1,0 @@
-org.xwiki.requiredrights.RequiredRightsScriptService
