+++ xwiki-platform-core/xwiki-platform-security/xwiki-platform-security-requiredrights/pom.xml
@@ -18,22 +18,29 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
 -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.xwiki.platform</groupId>
     <artifactId>xwiki-platform-core</artifactId>
-    <version>15.7-SNAPSHOT</version>
+    <version>15.6-SNAPSHOT</version>
   </parent>
-  <artifactId>xwiki-platform-security</artifactId>
-  <name>XWiki Platform - Security - Parent POM</name>
-  <description>Security APIs (Authentication, Authorization, and more)</description>
+  <artifactId>xwiki-platform-security-requiredrights</artifactId>
+  <name>XWiki Platform - Security - Required Rights - Parent POM</name>
   <packaging>pom</packaging>
-  <!-- Sorted Alphabetically -->
+  <description>Parent POM for the extenions related to Required Rights.</description>
   <modules>
-    <module>xwiki-platform-security-api</module>
-    <module>xwiki-platform-security-authentication</module>
-    <module>xwiki-platform-security-authorization</module>
-    <module>xwiki-platform-security-requiredrights</module>
+    <!-- Sorted Alphabetically -->
+    <module>xwiki-platform-security-requiredrights-api</module>
+    <module>xwiki-platform-security-requiredrights-ui</module>
   </modules>
+  <profiles>
+    <profile>
+      <id>integration-tests</id>
+      <modules>
+        <module>xwiki-platform-security-requiredrights-test</module>
+      </modules>
+    </profile>
+  </profiles>
 </project>
