+++ xwiki-platform-core/xwiki-platform-security/xwiki-platform-security-requiredrights/xwiki-platform-security-requiredrights-api/src/main/resources/META-INF/components.txt
@@ -1,0 +1,1 @@
+org.xwiki.security.requiredrights.RequiredRightsScriptService
