+++ xwiki-platform-core/xwiki-platform-security/xwiki-platform-security-requiredrights/xwiki-platform-security-requiredrights-test/pom.xml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+ * See the NOTICE file distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+-->
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.xwiki.platform</groupId>
+    <artifactId>xwiki-platform-security-requiredrights</artifactId>
+    <version>15.6-SNAPSHOT</version>
+  </parent>
+  <artifactId>xwiki-platform-security-requiredrights-test</artifactId>
+  <name>XWiki Platform - Security - Required Rights - Tests - Parent POM</name>
+  <packaging>pom</packaging>
+  <description>XWiki Platform - Security - Required Rights - Tests - Parent POM</description>
+    <properties>
+        <!-- Don't run backward-compatibility checks in test modules since we don't consider them as public APIs -->
+        <xwiki.revapi.skip>true</xwiki.revapi.skip>
+        <!-- Don't run Checkstyle in test modules -->
+        <xwiki.checkstyle.skip>true</xwiki.checkstyle.skip>
+    </properties>
+    <modules>
+        <module>xwiki-platform-security-requiredrights-test-pageobjects</module>
+    </modules>
+    <profiles>
+        <profile>
+            <id>docker</id>
+            <modules>
+                <module>xwiki-platform-security-requiredrights-test-docker</module>
+            </modules>
+        </profile>
+    </profiles>
+</project>
